.claude/skills/miro-flow-maker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "miro_flow_maker"
version = "0.1.0"
description = "Miro board flow maker for AIPO system"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
miro-flow-maker = "miro_flow_maker.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["scripts/miro_flow_maker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
